swap problem and solution file for exercise-antipatterns
